Update how validity of values for parameters is checked. This change may alter results!
First, in COIN-OR code, isVal equates to CoinAbsFltEq in
CoinFloatEqual.hpp, which uses val1 == val2 (if val1 - val2) < eps,
while were using <= in isVal. This may cause lots of changes!

Second, we were not checking parameter range validity correctly. The
problem is that something like 1e13 and 1e13-1 are equal due to floating
point precision. We do not solve _that_ problem, but we avoid it in the
check() / val_is_allowed() methods for the parameters.